Dangerous change: completely disable python codecs and feed utf8 into the parser directly
You'll need to run "make parser" again now, as the parser changed.
A nice excuse for this change is that it will make things faster, but the
real reason is that I couldn't get it to work with the python codecs.
Another advantage is that we should now be able to get rid of those
encoding python scripts.